+++ src/xpress_compress.c
@@ -903,7 +903,7 @@ xpress_find_matches(struct xpress_compressor * restrict c,
 	const u8 *in_next = in_base;
 	const u8 * const in_end = in_base + in_nbytes;
 	struct lz_match *cache_ptr = c->match_cache;
-	unsigned long prev_hash = 0;
+	u32 prev_hash = 0;
 
 	bt_matchfinder_init(&c->bt_mf);
